api/requirements.txt
biopython==1.80
fastapi==0.89.1
pysam==0.19.0
requests==2.26.0
uvicorn==0.20.0
pandas==1.5.3
